feat: 공통 로직 ITCrawler로 이름 변경 및 공통 모듈 추가
