+++ src/plugins/common/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod it_crawler;
