+++ src/plugins/cse/employment.rs
@@ -3,10 +3,10 @@ use crate::{
     core::{SsufidPlugin, SsufidPost},
 };
 
-use super::CseCrawler;
+use crate::plugins::common::it_crawler::ITCrawler;
 
 pub struct CseEmploymentPlugin {
-    crawler: CseCrawler<Self>,
+    crawler: ITCrawler<Self>,
 }
 
 impl SsufidPlugin for CseEmploymentPlugin {
@@ -27,9 +27,9 @@ impl Default for CseEmploymentPlugin {
 }
 
 impl CseEmploymentPlugin {
-    pub fn new() -> Self {
+    fn new() -> Self {
         Self {
-            crawler: CseCrawler::new(),
+            crawler: ITCrawler::new(),
         }
     }
 }
