+++ src/plugins/mod.rs
@@ -2,3 +2,5 @@ pub mod cse;
 pub mod media;
 pub mod ssu_catch;
 pub mod ssu_path;
+
+mod common;
